fix warnings under MinGW
The first problem is caused by the attribute not being matched properly
(gnu_printf vs printf). Simplest solution is to use the internal macro.

Second and third are wrong formats under Windows. For the third,
instead of using snprintf to write to a char buffer, it's simpler to
just use std::to_string.

+++ src/dispatcher/miniserver.cpp
@@ -583,7 +583,7 @@ static int get_port(
         *port = ntohs(reinterpret_cast<struct sockaddr_in6*>(&sockinfo)->sin6_port);
     }
     UpnpPrintf(UPNP_INFO, MSERV, __FILE__, __LINE__,
-               "sockfd = %d, .... port = %d\n", sockfd, static_cast<int>(*port));
+               "sockfd = %d, .... port = %d\n", static_cast<int>(sockfd), static_cast<int>(*port));
 
     return 0;
 }
